Split hook and JSX component
