+++ libs/i18n/locales/en/translation.json
@@ -262,6 +262,11 @@
   "Device name": "Device name",
   "Device enrollment request": "Device enrollment request",
   "Enrollment request is already approved.": "Enrollment request is already approved.",
+  "Add labels to select a fleet.": "Add labels to select a fleet.",
+  "Fleet label match cannot cannot be evaluated when there are invalid labels.": "Fleet label match cannot cannot be evaluated when there are invalid labels.",
+  "No fleet is matching the selected labels.": "No fleet is matching the selected labels.",
+  "More than one fleet is matching the selected labels. The device will ignore the fleets' configurations.": "More than one fleet is matching the selected labels. The device will ignore the fleets' configurations.",
+  "Check for matching fleet(s) failed. {{errorMessage}}": "Check for matching fleet(s) failed. {{errorMessage}}",
   "Delete {{resourceType}} ?": "Delete {{resourceType}} ?",
   "Are you sure you want to delete {resourceType} <3>{resourceName}</3>?": "Are you sure you want to delete {resourceType} <3>{resourceName}</3>?",
   "Saving...": "Saving...",
@@ -384,11 +389,6 @@
   "Light": "Light",
   "Dark": "Dark",
   "Theme": "Theme",
-  "Add labels to select a fleet.": "Add labels to select a fleet.",
-  "Fleet label match cannot cannot be evaluated when there are invalid labels.": "Fleet label match cannot cannot be evaluated when there are invalid labels.",
-  "No fleet is matching the selected labels.": "No fleet is matching the selected labels.",
-  "More than one fleet is matching the selected labels. The device will ignore the fleets' configurations.": "More than one fleet is matching the selected labels. The device will ignore the fleets' configurations.",
-  "Check for matching fleet(s) failed. {{errorMessage}}": "Check for matching fleet(s) failed. {{errorMessage}}",
   "Websocket error occured": "Websocket error occured",
   "{{count}} years ago_one": "{{count}} year ago",
   "{{count}} years ago_other": "{{count}} years ago",
